rename class: FiniteTransitionSystem -> TransitionSystem
because "finite" is ambiguous and elongates the class name, rendering it less readable. The `transys` package is intended to provide facilities for automata-theoretic methods, which typically involve graph representations with a finite number of nodes. Transition systems with an infinite number of states are recursively defined by classes in `tulip.hybrid` and elsewhere. Therefore namespaces serve the purpose of distinguishing between the two. In that sense, "finite" is superfluous.