add class: SystemGraph, replacement for LabeledDiGraph
It is called "system" instead of "rooted" graph because "initial_nodes" is more readable in the context of automata-theoretic methods compared to "root_nodes".

It includes methods for:

- checking consistency
- enforcing consistency (where possible)
- dump using GraphViz

`tikz` and other wild formats are removed from this base class. Variety can be provided as export functions, to keep the core simple and maintainable.

`dumps_dot` renamed to `to_pydot`. Subclasses are supposed to override this, in order to define their own style. No more `_save` lurking around.